chore: translations fallback (#12136)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->

<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces a fallback mechanism for translations in the
`LanguageProvider`. If a translation is unavailable in the localized
bundle, it will now fallback to a complete translations file.

### Detailed summary
- Updated `LanguageProvider` in `packages/localization/src/Provider.tsx`
to include a fallback for missing translations.
- If a translation key is not found, it now checks the `full`
translations file before defaulting to the key itself.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->