fix: Crash when chain query aptos in web (#12438)
page crashes:

https://pancakeswap.finance/swap?chain=aptos


<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->

<!-- start pr-codex -->

---

## PR-Codex overview
This PR introduces support for the `Aptos` blockchain by adding a new
function and updating relevant files to accommodate this change. It
modifies network selection and error handling to reflect the differences
in how `Aptos` is treated compared to other chains.

### Detailed summary
- Added `isAptos` function in `packages/chains/src/utils.ts` to check if
a chain ID corresponds to `Aptos`.
- Updated `apps/web/src/components/SearchModal/SwapNetworkSelection.tsx`
to correct the spelling of `Aptos` and ensure it is not shown in the
network selector.
- Modified `apps/solana/src/components/SentryErrorBoundary.tsx` to
change the reload behavior to navigate to a specific URL.
- Updated `apps/web/src/wallet/util/getQueryChainId.ts` to handle
`Aptos` by returning `undefined` if the chain is not supported.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->