fix: Crash when calculating mid price in swap (#11240)
<!--
Before opening a pull request, please read the [contributing
guidelines](https://github.com/pancakeswap/pancake-frontend/blob/develop/CONTRIBUTING.md)
first
-->

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on enhancing error handling in the `exchange.ts` file by
wrapping the calculation of `midPrice` and the subsequent output amount
in a `try-catch` block. This ensures that if an error occurs during the
calculation, it will be logged, and a default value will be assigned.

### Detailed summary
- Added a `try-catch` block around the calculation of `midPrice`.
- Introduced error logging with `console.error` for better debugging.
- Set `outputAmountWithoutPriceImpact` to a default value of
`CurrencyAmount.fromRawAmount(trade.outputAmount.currency, 0)` in case
of an error.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->